Add support for very large altitudes into render.frag
NOTE: Eclipse precomputers still don't support it. They'll need to use
some other texture coordinate transformations for the cases when
altitude>atmosphereHeight. Otherwise, if we simply project the camera
rays to the TOA, we'll get wrong results due to solar elevation being
the same for all of the projected points.