Redo progress reporting to avoid triggering recursive repaints
Now it's the responsibility of the caller code of AtmosphereRenderer to
make sure the renderer is initialized with whatever number of steps it
needs. What the caller gets is the ability to report status outside the
rendering process, thus breaking the painting recursion when the GL
widget and the GUI are in the same window.