Prevent git "unsafe directory" error when running make install as root
Installation via `make install` is normally done as root. The `install`
target depends on the build targets, which always try to update
`version.cpp` via `git-describe`. But git, since 2.35.2 (see [1] for
details), checks owner of the repository against current user, so we
either have to add a safe directory to the user's `.gitconfig`, or run
`git-describe` from the repo-owning user. This patch goes the latter way.

1: https://github.blog/2022-04-12-git-security-vulnerability-announced