Revert logarithmic scale for irradiance texture
This reverts commits dcc25bae51adcc76f1804f81249586e4eb9285ab and
192e7e95b988cee27f85d23ad7d4df4fcbc3d2de.

It appears that we do accumulate delta irradiance when computing
contributions of different scatterers. This breaks when the shader
writes logarithms of irradiance instead of actual irradiance, and
the blending engine can't natively compute
log(exp(framebuf)+exp(input)).

So, this issue has to be solved before this feature is re-introduced.