Fix guided interpolation algorithm
Interpolation angle was not computed correctly: when simply sampling
interpolation guides texture, we get an angle corresponding to the
middle of the space between detailed rows. But if the angle is far from
zero, then this sample is very wrong away from the middle. This patch
fixes the sampling by finding the correct guide sampling point using
binary search.