Add a workaround for poor NPOT mipmap implementations
We use the deepest mipmap level as the average of texture data in order
to compute integrals in EclipsedDoubleScatteringPrecomputer. Some OpenGL
implementations, however, appear to handle non-power-of-two textures
poorly when computing mipmap levels. The result is that, when a texture
dimension is close to a power of two but not exactly it, the whole mip
map displays ugly aliasing, and the deepest 1×1 level is very far from
the average of the texels of level 0.

To work around such OpenGL implementations, we resize the base level
texture to the closest smaller power-of-two texture, and compute the mip
map from it instead of the initial texture. The resize introduces a
small inaccuracy, but the result is still much higher quality than if we
used glGenerateMipmap() on the initial NPOT texture.

On good implementations such as NVIDIA GPU drivers, the workaround is
disabled.

Fixes #14