Don't embed Rayleigh phase function in the example's texture
When it's embedded, the resulting texture becomes badly conditioned for
interpolation. In particular, there appears a strange "cap" around the
zenith, where the interpolant is overly curved by bilinear filtering.

Computing the phase function on the fly, as will be the effect of this
commit, changes fullscreen ShowMySky performance on my Intel UHD
Graphics 620 from about 26 FPS to about 25.5 FPS, which corresponds to a
frame rate limit of 1326 FPS if this were the only thing computed.