Fix numeric instability in circlesIntersectionArea()
R1 and R2 are greater than d in the case being changed, and sometimes d
is *much* less than R1 and R2. In these cases adding R1 or R2 to it and
subtracting the R2 or R1 respectively results in catastrophic
cancellation, which produces visual artifacts particularly noticeable
when R1≈R2.